requirement_main.txt
openpyxl
line-bot-sdk
bs4
pandas
configparser
flask
#re
#json
pymysql